Fix Float Int Precision Issue

+++ gpu_check.py
@@ -349,8 +349,8 @@ def gpu_monitor(miner_id, DEBUG = False):
 						power_delta_dec = int(gpu.default_power_limit * 0.03)
 
 						# Is Over Limit Now! Reduce Power Immediately:
-						if gpu.power_limit > pw_limit_ub * float(gpu.default_power_limit):
-							new_power_limit = pw_limit_ub * float(gpu.default_power_limit)
+						if gpu.power_limit > int(pw_limit_ub * float(gpu.default_power_limit)):
+							new_power_limit = int(pw_limit_ub * float(gpu.default_power_limit))
 							process = subprocess.Popen("nvidia-smi.exe -i {} -pl {}".format(device_id, new_power_limit), stdout = subprocess.PIPE, stderr = subprocess.PIPE, shell=True)
 							output, error = process.communicate()
 							print("GPU #{}: Over Power Limit UB. Reset to UB: {}".format(gpu.gid, output))
